+++ contracts/library/ManagementLib.sol
@@ -5,9 +5,9 @@ library ManagementLib {
     function _hasDuplicates(
         address[] memory _addresses
     ) internal pure returns (bool) {
-        uint addressesLength = _addresses.length;
-        for (uint i = 0; i < addressesLength - 1; i++) {
-            for (uint j = i + 1; j < addressesLength; j++) {
+        uint256 addressesLength = _addresses.length;
+        for (uint256 i = 0; i < addressesLength - 1; i++) {
+            for (uint256 j = i + 1; j < addressesLength; j++) {
                 if (_addresses[i] == _addresses[j]) {
                     return true;
                 }
